pyproject.toml
[tool.poetry]
name = "keycloak-config-generator"
version = "0.1.0"
description = "A CLI tool to generate Keycloak realm configurations from YAML files."
authors = ["Your Name <you@example.com>"]
packages = [{include = "keycloak_generator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
PyYAML = "^6.0.1"
pydantic = "^2.11.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
ruff = "^0.0.292"

[tool.poetry.scripts]
kc-generate = "keycloak_generator.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104", "D107"]
